refactor for tests
